# swagger.yaml
openapi: 3.0.0
info:
  title: RESTFUL API for 2nd sem WebDev
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/
    description: Localhost development server 
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token

  schemas:
    Game:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        publisher:
          type: string
        genre:
          type: string
        details:
          type: string
        releaseYear:
          type: number
        recommended:
          type: boolean
    
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        date:
          type: date

paths:

  /user/register:
    post:
      summary: Register a new User
      description: Register a new User in the database
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"

      responses:
        201:
          description: User created succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  _id:
                    type: string

  /user/login:
    post:
      summary: Login a User
      description: Login a User in order to get JWT token (auth-token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      
      responses:
        200:
          description: User logged in succesfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  data:
                    type: string


  /games:
    get:
      summary: Retrieves a list of games
      description: Retrieves a list of games as JSON objects.
      responses:
        200:
          description: A list of Game JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
    
    post:
      summary: Create a new Game
      description: Create a new Game
      security:
        - ApiKeyAuth: []      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"

      responses:
        201:
          description: Game created succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
  
  /games/{id}:
    get:
      summary: Specific Game
      description: Retrieves a specific Game based on it id.
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      responses:
        200:
          description: A Game in the format of a JSON object.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
   
    put:
      summary: Updates a specific Game
      description: Updates a specific Game based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"

      responses:
        201:
          description: Game updated succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"

    delete:
      summary: Deletes a specific Game
      description: Deletes a specific Game based on it id
      security:
        - ApiKeyAuth: []            
      parameters:
        - in: path
          name: id
          required: true
          description: MongoDB id
          schema:
            type: string      

      responses:
        201:
          description: Game deleted succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"

  


    

    